refactor(configurator): split boards page modules
